kvm: put functions shared by both drivers in kvm_common.c
