+++ libvmi/driver/kvm/kvm.c
@@ -88,14 +88,12 @@ kvm_get_memory_patch(
     addr_t paddr,
     uint32_t length)
 {
-    char *buf = g_try_malloc0(length + 1);
-    if ( !buf )
-        return NULL;
+    kvm_instance_t *kvm = kvm_get_instance(vmi);
 
     if (!kvm->kvmi_dom)
         return NULL;
 
-    buffer = g_malloc0(length);
+    char* buffer = g_malloc0(length);
     if (kvmi_read_physical(kvm->kvmi_dom, paddr, buffer, length) < 0) {
         g_free(buffer);
         return NULL;
@@ -471,125 +469,6 @@ kvm_destroy(
 
 }
 
-uint64_t
-kvm_get_id_from_name(
-    vmi_instance_t vmi,
-    const char *name)
-{
-    virDomainPtr dom = NULL;
-    uint64_t domainid = VMI_INVALID_DOMID;
-    kvm_instance_t *kvm = kvm_get_instance(vmi);
-
-    dom = kvm->libvirt.virDomainLookupByName(kvm->conn, name);
-    if (NULL == dom) {
-        dbprint(VMI_DEBUG_KVM, "--failed to find kvm domain\n");
-        domainid = VMI_INVALID_DOMID;
-    } else {
-
-        domainid = (uint64_t) kvm->libvirt.virDomainGetID(dom);
-        if (domainid == (uint64_t)-1) {
-            dbprint(VMI_DEBUG_KVM, "--requested kvm domain may not be running\n");
-            domainid = VMI_INVALID_DOMID;
-        }
-    }
-
-    if (dom)
-        kvm->libvirt.virDomainFree(dom);
-
-    return domainid;
-}
-
-status_t
-kvm_get_name_from_id(
-    vmi_instance_t vmi,
-    uint64_t domainid,
-    char **name)
-{
-    virDomainPtr dom = NULL;
-    const char* temp_name = NULL;
-    kvm_instance_t *kvm = kvm_get_instance(vmi);
-
-    dom = kvm->libvirt.virDomainLookupByID(kvm->conn, domainid);
-    if (NULL == dom) {
-        dbprint(VMI_DEBUG_KVM, "--failed to find kvm domain\n");
-        return VMI_FAILURE;
-    }
-
-    temp_name = kvm->libvirt.virDomainGetName(dom);
-    *name = temp_name ? strdup(temp_name) : NULL;
-
-    if (dom)
-        kvm->libvirt.virDomainFree(dom);
-
-    if (*name) {
-        return VMI_SUCCESS;
-    }
-
-    return VMI_FAILURE;
-}
-
-uint64_t
-kvm_get_id(
-    vmi_instance_t vmi)
-{
-    return kvm_get_instance(vmi)->id;
-}
-
-void
-kvm_set_id(
-    vmi_instance_t vmi,
-    uint64_t domainid)
-{
-    kvm_get_instance(vmi)->id = domainid;
-}
-
-status_t
-kvm_check_id(
-    vmi_instance_t vmi,
-    uint64_t domainid)
-{
-    virDomainPtr dom = NULL;
-    kvm_instance_t *kvm = kvm_get_instance(vmi);
-
-    dom = kvm->libvirt.virDomainLookupByID(kvm->conn, domainid);
-    if (NULL == dom) {
-        dbprint(VMI_DEBUG_KVM, "--failed to find kvm domain\n");
-        return VMI_FAILURE;
-    }
-
-    if (dom)
-        kvm->libvirt.virDomainFree(dom);
-
-    return VMI_SUCCESS;
-}
-
-status_t
-kvm_get_name(
-    vmi_instance_t vmi,
-    char **name)
-{
-    kvm_instance_t *kvm = kvm_get_instance(vmi);
-
-    const char *tmpname = kvm->libvirt.virDomainGetName(kvm->dom);
-
-    // don't need to deallocate the name, it will go away with the domain object
-
-    if (NULL != tmpname) {
-        *name = strdup(tmpname);
-        return VMI_SUCCESS;
-    } else {
-        return VMI_FAILURE;
-    }
-}
-
-void
-kvm_set_name(
-    vmi_instance_t vmi,
-    const char *name)
-{
-    kvm_get_instance(vmi)->name = strndup(name, 500);
-}
-
 status_t
 kvm_get_memsize(
     vmi_instance_t vmi,
@@ -609,6 +488,7 @@ kvm_get_memsize(
 
     *allocated_ram_size = max_gfn * vmi->page_size;
     *maximum_physical_address = max_gfn << vmi->page_shift;
+    return VMI_SUCCESS;
 }
 
 status_t kvm_request_page_fault (
@@ -840,65 +720,6 @@ kvm_get_vcpureg(
     if (get_kvmi_registers(kvm_get_instance(vmi), reg, value))
         return VMI_SUCCESS;
 
-    return VMI_FAILURE;
-}
-
-void *
-kvm_read_page(
-    vmi_instance_t vmi,
-    addr_t page)
-{
-    addr_t paddr = page << vmi->page_shift;
-
-    return memory_cache_insert(vmi, paddr);
-}
-
-status_t
-kvm_write(
-    vmi_instance_t vmi,
-    addr_t paddr,
-    void *buf,
-    uint32_t length)
-{
-    return kvm_put_memory(vmi, paddr, length, buf);
-}
-
-int
-kvm_is_pv(
-    vmi_instance_t UNUSED(vmi))
-{
-    return 0;
-}
-
-status_t
-kvm_test(
-    uint64_t domainid,
-    const char *name,
-    uint64_t UNUSED(init_flags),
-    vmi_init_data_t* UNUSED(init_data))
-{
-    struct vmi_instance _vmi = {0};
-    vmi_instance_t vmi = &_vmi;
-
-    if ( VMI_FAILURE == kvm_init(vmi, 0, NULL) )
-        return VMI_FAILURE;
-
-    if (name) {
-        domainid = kvm_get_id_from_name(vmi, name);
-        if (domainid != VMI_INVALID_DOMID)
-            return VMI_SUCCESS;
-    }
-
-    if (domainid != VMI_INVALID_DOMID) {
-        char *_name = NULL;
-        status_t rc = kvm_get_name_from_id(vmi, domainid, &_name);
-        free(_name);
-
-        if ( VMI_SUCCESS == rc )
-            return rc;
-    }
-
-    kvm_destroy(vmi);
     return VMI_FAILURE;
 }
