+++ libvmi/driver/kvm/kvm_common.c
@@ -1,0 +1,208 @@
+/* The LibVMI Library is an introspection library that simplifies access to
+ * memory in a target virtual machine or in a file containing a dump of
+ * a system's physical memory.  LibVMI is based on the XenAccess Library.
+ *
+ * Copyright 2011 Sandia Corporation. Under the terms of Contract
+ * DE-AC04-94AL85000 with Sandia Corporation, the U.S. Government
+ * retains certain rights in this software.
+ *
+ * Author: Mathieu Tarral (mathieu.tarral@ssi.gouv.fr)
+ *
+ * This file is part of LibVMI.
+ *
+ * LibVMI is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the
+ * Free Software Foundation, either version 3 of the License, or (at your
+ * option) any later version.
+ *
+ * LibVMI is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public
+ * License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with LibVMI.  If not, see <http://www.gnu.org/licenses/>.
+ */
+#include "private.h"
+#include "driver/driver_wrapper.h"
+#include "driver/memory_cache.h"
+#include "driver/kvm/kvm.h"
+#include "driver/kvm/kvm_private.h"
+
+uint64_t
+kvm_get_id_from_name(
+    vmi_instance_t vmi,
+    const char *name)
+{
+    virDomainPtr dom = NULL;
+    uint64_t domainid = VMI_INVALID_DOMID;
+    kvm_instance_t *kvm = kvm_get_instance(vmi);
+
+    dom = kvm->libvirt.virDomainLookupByName(kvm->conn, name);
+    if (NULL == dom) {
+        dbprint(VMI_DEBUG_KVM, "--failed to find kvm domain\n");
+        domainid = VMI_INVALID_DOMID;
+    } else {
+
+        domainid = (uint64_t) kvm->libvirt.virDomainGetID(dom);
+        if (domainid == (uint64_t)-1) {
+            dbprint(VMI_DEBUG_KVM, "--requested kvm domain may not be running\n");
+            domainid = VMI_INVALID_DOMID;
+        }
+    }
+
+    if (dom)
+        kvm->libvirt.virDomainFree(dom);
+
+    return domainid;
+}
+
+status_t
+kvm_get_name_from_id(
+    vmi_instance_t vmi,
+    uint64_t domainid,
+    char **name)
+{
+    virDomainPtr dom = NULL;
+    const char* temp_name = NULL;
+    kvm_instance_t *kvm = kvm_get_instance(vmi);
+
+    dom = kvm->libvirt.virDomainLookupByID(kvm->conn, domainid);
+    if (NULL == dom) {
+        dbprint(VMI_DEBUG_KVM, "--failed to find kvm domain\n");
+        return VMI_FAILURE;
+    }
+
+    temp_name = kvm->libvirt.virDomainGetName(dom);
+    *name = temp_name ? strdup(temp_name) : NULL;
+
+    if (dom)
+        kvm->libvirt.virDomainFree(dom);
+
+    if (*name) {
+        return VMI_SUCCESS;
+    }
+
+    return VMI_FAILURE;
+}
+
+uint64_t
+kvm_get_id(
+    vmi_instance_t vmi)
+{
+    return kvm_get_instance(vmi)->id;
+}
+
+void
+kvm_set_id(
+    vmi_instance_t vmi,
+    uint64_t domainid)
+{
+    kvm_get_instance(vmi)->id = domainid;
+}
+
+status_t
+kvm_check_id(
+    vmi_instance_t vmi,
+    uint64_t domainid)
+{
+    virDomainPtr dom = NULL;
+    kvm_instance_t *kvm = kvm_get_instance(vmi);
+
+    dom = kvm->libvirt.virDomainLookupByID(kvm->conn, domainid);
+    if (NULL == dom) {
+        dbprint(VMI_DEBUG_KVM, "--failed to find kvm domain\n");
+        return VMI_FAILURE;
+    }
+
+    if (dom)
+        kvm->libvirt.virDomainFree(dom);
+
+    return VMI_SUCCESS;
+}
+
+status_t
+kvm_get_name(
+    vmi_instance_t vmi,
+    char **name)
+{
+    kvm_instance_t *kvm = kvm_get_instance(vmi);
+
+    const char *tmpname = kvm->libvirt.virDomainGetName(kvm->dom);
+
+    // don't need to deallocate the name, it will go away with the domain object
+
+    if (NULL != tmpname) {
+        *name = strdup(tmpname);
+        return VMI_SUCCESS;
+    } else {
+        return VMI_FAILURE;
+    }
+}
+
+void
+kvm_set_name(
+    vmi_instance_t vmi,
+    const char *name)
+{
+    kvm_get_instance(vmi)->name = strndup(name, 500);
+}
+
+void *
+kvm_read_page(
+    vmi_instance_t vmi,
+    addr_t page)
+{
+    addr_t paddr = page << vmi->page_shift;
+
+    return memory_cache_insert(vmi, paddr);
+}
+
+status_t
+kvm_write(
+    vmi_instance_t vmi,
+    addr_t paddr,
+    void *buf,
+    uint32_t length)
+{
+    return kvm_put_memory(vmi, paddr, length, buf);
+}
+
+int
+kvm_is_pv(
+    vmi_instance_t UNUSED(vmi))
+{
+    return 0;
+}
+
+status_t
+kvm_test(
+    uint64_t domainid,
+    const char *name,
+    uint64_t UNUSED(init_flags),
+    vmi_init_data_t* UNUSED(init_data))
+{
+    struct vmi_instance _vmi = {0};
+    vmi_instance_t vmi = &_vmi;
+
+    if ( VMI_FAILURE == kvm_init(vmi, 0, NULL) )
+        return VMI_FAILURE;
+
+    if (name) {
+        domainid = kvm_get_id_from_name(vmi, name);
+        if (domainid != VMI_INVALID_DOMID)
+            return VMI_SUCCESS;
+    }
+
+    if (domainid != VMI_INVALID_DOMID) {
+        char *_name = NULL;
+        status_t rc = kvm_get_name_from_id(vmi, domainid, &_name);
+        free(_name);
+
+        if ( VMI_SUCCESS == rc )
+            return rc;
+    }
+
+    kvm_destroy(vmi);
+    return VMI_FAILURE;
+}
